[REFACTOR] add components at MyPageLobby
