+++ src/app/ChannelLobby/button/leftSideButtons.tsx
@@ -1,10 +1,6 @@
 'use client';
 
-import type { NextPage } from "next";
 import { useState, useCallback } from "react";
-import MyPageFrame from "@/components/my-page-frame";
-import PortalPopup from "@/components/portal-popup";
-
 
 export default function LeftSideButtons() {
     const [isMyPageFrameOpen, setMyPageFrameOpen] = useState(false);
@@ -41,15 +37,6 @@ export default function LeftSideButtons() {
             src="/sidebutton21.svg"
             onClick={onSideButton21Click}
         />
-        {isMyPageFrameOpen && (
-            <PortalPopup
-                overlayColor="rgba(113, 113, 113, 0.3)"
-                placement="Centered"
-                onOutsideClick={closeMyPageFrame}
-            >
-                <MyPageFrame onClose={closeMyPageFrame} />
-            </PortalPopup>
-        )}
     </>
     );
 }
