+++ src/app/GameLobby/button/leftSideButtons.tsx
@@ -1,9 +1,6 @@
 'use client';
 
-import type { NextPage } from "next";
 import { useState, useCallback } from "react";
-import MyPageFrame from "@/components/my-page-frame";
-import PortalPopup from "@/components/portal-popup";
 
 
 export default function LeftSideButtons() {
@@ -41,15 +38,6 @@ export default function LeftSideButtons() {
             src="/sidebutton21.svg"
             onClick={onSideButton21Click}
         />
-        {isMyPageFrameOpen && (
-            <PortalPopup
-                overlayColor="rgba(113, 113, 113, 0.3)"
-                placement="Centered"
-                onOutsideClick={closeMyPageFrame}
-            >
-                <MyPageFrame onClose={closeMyPageFrame} />
-            </PortalPopup>
-        )}
     </>
     );
 }
