+++ src/app/MyPageLobby/match/matchCard.tsx
@@ -1,0 +1,53 @@
+
+
+
+export default function MatchCard() {
+    return (
+        <div className="relative w-[56.25rem] h-[4.06rem]">
+        <img
+          className="absolute h-[45%] w-[4%] top-[20.96%] right-[71.44%] bottom-[34.04%] left-[24.56%] max-w-full overflow-hidden max-h-full opacity-[0.65]"
+          alt=""
+          src="/winnercrown.svg"
+        />
+        <div className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_420px)] w-[37.75rem] h-[2.31rem] text-[1.75rem]">
+          <div className="absolute top-[0rem] left-[20.69rem] w-[13.75rem] flex flex-row items-center justify-start gap-[0.63rem]">
+            <img
+              className="relative w-[2rem] h-[2rem] object-cover"
+              alt=""
+              src="/pngegg-4@2x.png"
+            />
+            <div className="flex flex-row items-end justify-center gap-[0.63rem]">
+              <i className="relative">asdf</i>
+              <div className="relative text-[0.88rem]">vs 75.0%</div>
+            </div>
+          </div>
+          <i className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_40px)] inline-block text-center w-[4.31rem] h-[2.31rem]">
+            5 : 4
+          </i>
+          <div className="absolute top-[calc(50%_-_18.5px)] left-[0rem] flex flex-row items-center justify-end gap-[0.63rem] text-right">
+            <i className="relative inline-block font-bold w-[13.75rem] h-[2.31rem] shrink-0">
+              testNickname
+            </i>
+            <img
+              className="relative w-[2rem] h-[2rem] object-cover"
+              alt=""
+              src="/pngegg-3@2x.png"
+            />
+          </div>
+        </div>
+        <div className="absolute top-[2.31rem] left-[46rem] text-[0.88rem] w-[10rem] h-[1.1rem] text-right">
+          2023.07.26 17:09:34
+        </div>
+        <div className="absolute top-[0.88rem] left-[41rem] text-[1rem] w-[15rem] h-[1.19rem] text-right">
+          1 ball, speed moderate
+        </div>
+        <img 
+          className="absolute top-[3.94rem] left-[0rem] w-[56.25rem] h-[0.13rem]"
+          alt=""
+          src="/line-14.svg"
+        /> {/* 이거 마지막은 라인 없어야함, 근데 그림자 효과줘서 구분하는게 더 좋지 않나?  */}
+        <div className="absolute top-[0rem] left-[0.19rem]">{` `}</div>
+      </div>
+    );
+}
+
