--- src/components/my-page-frame.tsx
@@ -1,817 +1,0 @@
-import type { NextPage } from "next";
-
-type MyPageFrameType = {
-  onClose?: () => void;
-};
-
-const MyPageFrame: NextPage<MyPageFrameType> = ({ onClose }) => {
-  return (
-    <div className="relative w-[62.5rem] h-[40.63rem] max-w-full max-h-full overflow-auto text-left text-[1rem] text-dimgray font-inria-sans">
-      <div className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] [background:linear-gradient(240.36deg,_#f5f5f5,_#fafbff)]">
-        <div className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] rounded-3xs box-border border-[3px] border-solid border-dimgray" />
-        <div className="absolute h-[58.15%] w-[94.4%] top-[38.15%] right-[2.8%] bottom-[3.69%] left-[2.8%] rounded-3xs bg-gray shadow-[0px_2px_10px_rgba(0,_0,_0,_0.25)]" />
-        <div className="absolute top-[calc(50%_-_69px)] left-[calc(50%_-_450px)] h-[22.5rem] overflow-y-auto flex flex-col items-start justify-start gap-[0.63rem]">
-          <div className="relative w-[56.25rem] h-[4.06rem]">
-            <img
-              className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <img
-              className="absolute h-[45%] w-[4%] top-[20.96%] right-[71.44%] bottom-[34.04%] left-[24.56%] max-w-full overflow-hidden max-h-full opacity-[0.65]"
-              alt=""
-              src="/winnercrown.svg"
-            />
-            <div className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_420px)] w-[37.75rem] h-[2.31rem] text-[1.75rem]">
-              <div className="absolute top-[0rem] left-[20.69rem] w-[13.75rem] flex flex-row items-center justify-start gap-[0.63rem]">
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-4@2x.png"
-                />
-                <div className="flex flex-row items-end justify-center gap-[0.63rem]">
-                  <i className="relative">aaaaaaaaaa</i>
-                  <div className="relative text-[0.88rem]">vs 75.0%</div>
-                </div>
-              </div>
-              <i className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_40px)] inline-block text-center w-[4.31rem] h-[2.31rem]">
-                5 : 2
-              </i>
-              <div className="absolute top-[calc(50%_-_18.5px)] left-[0rem] flex flex-row items-center justify-end gap-[0.63rem] text-right">
-                <i className="relative inline-block font-bold w-[13.75rem] h-[2.31rem] shrink-0">
-                  testNickname
-                </i>
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-3@2x.png"
-                />
-              </div>
-            </div>
-            <div className="absolute top-[2.31rem] left-[49.06rem] text-[0.88rem] text-right">
-              2023.07.26 17:09:34
-            </div>
-            <div className="absolute top-[0.88rem] left-[46.69rem] text-right">
-              1 ball, speed moderate
-            </div>
-            <img
-              className="absolute top-[3.94rem] left-[0rem] w-[56.25rem] h-[0.13rem]"
-              alt=""
-              src="/line-14.svg"
-            />
-            <div className="absolute top-[0rem] left-[0.19rem]">{` `}</div>
-          </div>
-          <div className="relative w-[56.25rem] h-[4.06rem]">
-            <img
-              className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <img
-              className="absolute h-[45%] w-[4%] top-[20.96%] right-[52%] bottom-[34.04%] left-[44%] max-w-full overflow-hidden max-h-full opacity-[0.65]"
-              alt=""
-              src="/winnercrown.svg"
-            />
-            <div className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_420px)] w-[37.75rem] h-[2.31rem] text-[1.75rem]">
-              <div className="absolute top-[0rem] left-[20.69rem] w-[13.75rem] flex flex-row items-center justify-start gap-[0.63rem]">
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-4@2x.png"
-                />
-                <div className="flex flex-row items-end justify-center gap-[0.63rem]">
-                  <i className="relative font-bold">aaaaa</i>
-                  <div className="relative text-[0.88rem]">vs 75.0%</div>
-                </div>
-              </div>
-              <i className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_40px)] inline-block text-center w-[4.31rem] h-[2.31rem]">
-                2 : 5
-              </i>
-              <div className="absolute top-[calc(50%_-_18.5px)] left-[0rem] flex flex-row items-center justify-end gap-[0.63rem] text-right">
-                <i className="relative inline-block w-[13.75rem] h-[2.31rem] shrink-0">
-                  testNickname
-                </i>
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-3@2x.png"
-                />
-              </div>
-            </div>
-            <div className="absolute top-[2.31rem] left-[49.06rem] text-[0.88rem] text-right">
-              2023.07.26 17:09:34
-            </div>
-            <div className="absolute top-[0.88rem] left-[46.69rem] text-right">
-              1 ball, speed moderate
-            </div>
-            <img
-              className="absolute top-[3.94rem] left-[0rem] w-[56.25rem] h-[0.13rem]"
-              alt=""
-              src="/line-14.svg"
-            />
-            <div className="absolute top-[0rem] left-[0.19rem]">custom</div>
-          </div>
-          <div className="relative w-[56.25rem] h-[4.06rem]">
-            <img
-              className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <img
-              className="absolute h-[45%] w-[4%] top-[20.96%] right-[52%] bottom-[34.04%] left-[44%] max-w-full overflow-hidden max-h-full opacity-[0.65]"
-              alt=""
-              src="/winnercrown.svg"
-            />
-            <div className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_420px)] w-[37.75rem] h-[2.31rem] text-[1.75rem]">
-              <div className="absolute top-[0rem] left-[20.69rem] w-[13.75rem] flex flex-row items-center justify-start gap-[0.63rem]">
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-4@2x.png"
-                />
-                <div className="flex flex-row items-end justify-center gap-[0.63rem]">
-                  <i className="relative font-bold">tjo</i>
-                  <div className="relative text-[0.88rem]">vs 75.0%</div>
-                </div>
-              </div>
-              <i className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_40px)] inline-block text-center w-[4.31rem] h-[2.31rem]">
-                3 : 5
-              </i>
-              <div className="absolute top-[calc(50%_-_18.5px)] left-[0rem] flex flex-row items-center justify-end gap-[0.63rem] text-right">
-                <i className="relative inline-block w-[13.75rem] h-[2.31rem] shrink-0">
-                  testNickname
-                </i>
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-3@2x.png"
-                />
-              </div>
-            </div>
-            <div className="absolute top-[2.31rem] left-[49.06rem] text-[0.88rem] text-right">
-              2023.07.26 17:09:34
-            </div>
-            <div className="absolute top-[0.88rem] left-[46.69rem] text-right">
-              1 ball, speed moderate
-            </div>
-            <img
-              className="absolute top-[3.94rem] left-[0rem] w-[56.25rem] h-[0.13rem]"
-              alt=""
-              src="/line-14.svg"
-            />
-            <div className="absolute top-[0rem] left-[0.19rem]">custom</div>
-          </div>
-          <div className="relative w-[56.25rem] h-[4.06rem]">
-            <img
-              className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <img
-              className="absolute h-[45%] w-[4%] top-[20.96%] right-[71.44%] bottom-[34.04%] left-[24.56%] max-w-full overflow-hidden max-h-full opacity-[0.65]"
-              alt=""
-              src="/winnercrown.svg"
-            />
-            <div className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_420px)] w-[37.75rem] h-[2.31rem] text-[1.75rem]">
-              <div className="absolute top-[0rem] left-[20.69rem] w-[13.75rem] flex flex-row items-center justify-start gap-[0.63rem]">
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-4@2x.png"
-                />
-                <div className="flex flex-row items-end justify-center gap-[0.63rem]">
-                  <i className="relative">sans</i>
-                  <div className="relative text-[0.88rem]">vs 75.0%</div>
-                </div>
-              </div>
-              <i className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_40px)] inline-block text-center w-[4.31rem] h-[2.31rem]">
-                5 : 4
-              </i>
-              <div className="absolute top-[calc(50%_-_18.5px)] left-[0rem] flex flex-row items-center justify-end gap-[0.63rem] text-right">
-                <i className="relative inline-block font-bold w-[13.75rem] h-[2.31rem] shrink-0">
-                  testNickname
-                </i>
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-3@2x.png"
-                />
-              </div>
-            </div>
-            <div className="absolute top-[2.31rem] left-[49.06rem] text-[0.88rem] text-right">
-              2023.07.26 17:09:34
-            </div>
-            <div className="absolute top-[0.88rem] left-[46.69rem] text-right">
-              1 ball, speed moderate
-            </div>
-            <img
-              className="absolute top-[3.94rem] left-[0rem] w-[56.25rem] h-[0.13rem]"
-              alt=""
-              src="/line-14.svg"
-            />
-            <div className="absolute top-[0rem] left-[0.19rem]">{` `}</div>
-          </div>
-          <div className="relative w-[56.25rem] h-[4.06rem]">
-            <img
-              className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <img
-              className="absolute h-[45%] w-[4%] top-[20.96%] right-[71.44%] bottom-[34.04%] left-[24.56%] max-w-full overflow-hidden max-h-full opacity-[0.65]"
-              alt=""
-              src="/winnercrown.svg"
-            />
-            <div className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_420px)] w-[37.75rem] h-[2.31rem] text-[1.75rem]">
-              <div className="absolute top-[0rem] left-[20.69rem] w-[13.75rem] flex flex-row items-center justify-start gap-[0.63rem]">
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-4@2x.png"
-                />
-                <div className="flex flex-row items-end justify-center gap-[0.63rem]">
-                  <i className="relative">asdf</i>
-                  <div className="relative text-[0.88rem]">vs 75.0%</div>
-                </div>
-              </div>
-              <i className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_40px)] inline-block text-center w-[4.31rem] h-[2.31rem]">
-                5 : 4
-              </i>
-              <div className="absolute top-[calc(50%_-_18.5px)] left-[0rem] flex flex-row items-center justify-end gap-[0.63rem] text-right">
-                <i className="relative inline-block font-bold w-[13.75rem] h-[2.31rem] shrink-0">
-                  testNickname
-                </i>
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-3@2x.png"
-                />
-              </div>
-            </div>
-            <div className="absolute top-[2.31rem] left-[49.06rem] text-[0.88rem] text-right">
-              2023.07.26 17:09:34
-            </div>
-            <div className="absolute top-[0.88rem] left-[46.69rem] text-right">
-              1 ball, speed moderate
-            </div>
-            <img
-              className="absolute top-[4rem] left-[0rem] w-[56.25rem] h-[0rem]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[0rem] left-[0.19rem]">{` `}</div>
-          </div>
-          <div className="relative w-[56.25rem] h-[4.06rem]">
-            <img
-              className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <img
-              className="absolute h-[45%] w-[4%] top-[20.96%] right-[71.44%] bottom-[34.04%] left-[24.56%] max-w-full overflow-hidden max-h-full opacity-[0.65]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_420px)] w-[37.75rem] h-[2.31rem] text-[1.75rem]">
-              <div className="absolute top-[0rem] left-[20.69rem] w-[13.75rem] flex flex-row items-center justify-start gap-[0.63rem]">
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-                <div className="flex flex-row items-end justify-center gap-[0.63rem]">
-                  <i className="relative">userName2</i>
-                  <div className="relative text-[0.88rem]">vs 75.0%</div>
-                </div>
-              </div>
-              <i className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_40px)] inline-block text-center w-[4.31rem] h-[2.31rem]">
-                5 : 4
-              </i>
-              <div className="absolute top-[calc(50%_-_18.5px)] left-[0rem] flex flex-row items-center justify-end gap-[0.63rem] text-right">
-                <i className="relative inline-block font-bold w-[13.75rem] h-[2.31rem] shrink-0">
-                  userName1
-                </i>
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-              </div>
-            </div>
-            <div className="absolute top-[2.31rem] left-[49.06rem] text-[0.88rem] text-right">
-              2023.07.26 17:09:34
-            </div>
-            <div className="absolute top-[0.88rem] left-[46.69rem] text-right">
-              1 ball, speed moderate
-            </div>
-            <img
-              className="absolute top-[4rem] left-[0rem] w-[56.25rem] h-[0rem]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[0rem] left-[0.19rem]">custom</div>
-          </div>
-          <div className="relative w-[56.25rem] h-[4.06rem]">
-            <img
-              className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <img
-              className="absolute h-[45%] w-[4%] top-[20.96%] right-[71.44%] bottom-[34.04%] left-[24.56%] max-w-full overflow-hidden max-h-full opacity-[0.65]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_420px)] w-[37.75rem] h-[2.31rem] text-[1.75rem]">
-              <div className="absolute top-[0rem] left-[20.69rem] w-[13.75rem] flex flex-row items-center justify-start gap-[0.63rem]">
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-                <div className="flex flex-row items-end justify-center gap-[0.63rem]">
-                  <i className="relative">aaaaaaaaaa</i>
-                  <div className="relative text-[0.88rem]">vs 75.0%</div>
-                </div>
-              </div>
-              <i className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_40px)] inline-block text-center w-[4.31rem] h-[2.31rem]">
-                5 : 2
-              </i>
-              <div className="absolute top-[calc(50%_-_18.5px)] left-[0rem] flex flex-row items-center justify-end gap-[0.63rem] text-right">
-                <i className="relative inline-block font-bold w-[13.75rem] h-[2.31rem] shrink-0">
-                  testNickname
-                </i>
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-              </div>
-            </div>
-            <div className="absolute top-[2.31rem] left-[49.06rem] text-[0.88rem] text-right">
-              2023.07.26 17:09:34
-            </div>
-            <div className="absolute top-[0.88rem] left-[46.69rem] text-right">
-              1 ball, speed moderate
-            </div>
-            <img
-              className="absolute top-[4rem] left-[0rem] w-[56.25rem] h-[0rem]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[0rem] left-[0.19rem]">{` `}</div>
-          </div>
-          <div className="relative w-[56.25rem] h-[4.06rem]">
-            <img
-              className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <img
-              className="absolute h-[45%] w-[4%] top-[20.96%] right-[52%] bottom-[34.04%] left-[44%] max-w-full overflow-hidden max-h-full opacity-[0.65]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_420px)] w-[37.75rem] h-[2.31rem] text-[1.75rem]">
-              <div className="absolute top-[0rem] left-[20.69rem] w-[13.75rem] flex flex-row items-center justify-start gap-[0.63rem]">
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-                <div className="flex flex-row items-end justify-center gap-[0.63rem]">
-                  <i className="relative font-bold">aaaaa</i>
-                  <div className="relative text-[0.88rem]">vs 75.0%</div>
-                </div>
-              </div>
-              <i className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_40px)] inline-block text-center w-[4.31rem] h-[2.31rem]">
-                2 : 5
-              </i>
-              <div className="absolute top-[calc(50%_-_18.5px)] left-[0rem] flex flex-row items-center justify-end gap-[0.63rem] text-right">
-                <i className="relative inline-block w-[13.75rem] h-[2.31rem] shrink-0">
-                  testNickname
-                </i>
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-              </div>
-            </div>
-            <div className="absolute top-[2.31rem] left-[49.06rem] text-[0.88rem] text-right">
-              2023.07.26 17:09:34
-            </div>
-            <div className="absolute top-[0.88rem] left-[46.69rem] text-right">
-              1 ball, speed moderate
-            </div>
-            <img
-              className="absolute top-[4rem] left-[0rem] w-[56.25rem] h-[0rem]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[0rem] left-[0.19rem]">custom</div>
-          </div>
-          <div className="relative w-[56.25rem] h-[4.06rem]">
-            <img
-              className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <img
-              className="absolute h-[45%] w-[4%] top-[20.96%] right-[52%] bottom-[34.04%] left-[44%] max-w-full overflow-hidden max-h-full opacity-[0.65]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_420px)] w-[37.75rem] h-[2.31rem] text-[1.75rem]">
-              <div className="absolute top-[0rem] left-[20.69rem] w-[13.75rem] flex flex-row items-center justify-start gap-[0.63rem]">
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-                <div className="flex flex-row items-end justify-center gap-[0.63rem]">
-                  <i className="relative font-bold">tjo</i>
-                  <div className="relative text-[0.88rem]">vs 75.0%</div>
-                </div>
-              </div>
-              <i className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_40px)] inline-block text-center w-[4.31rem] h-[2.31rem]">
-                3 : 5
-              </i>
-              <div className="absolute top-[calc(50%_-_18.5px)] left-[0rem] flex flex-row items-center justify-end gap-[0.63rem] text-right">
-                <i className="relative inline-block w-[13.75rem] h-[2.31rem] shrink-0">
-                  testNickname
-                </i>
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-              </div>
-            </div>
-            <div className="absolute top-[2.31rem] left-[49.06rem] text-[0.88rem] text-right">
-              2023.07.26 17:09:34
-            </div>
-            <div className="absolute top-[0.88rem] left-[46.69rem] text-right">
-              1 ball, speed moderate
-            </div>
-            <img
-              className="absolute top-[4rem] left-[0rem] w-[56.25rem] h-[0rem]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[0rem] left-[0.19rem]">custom</div>
-          </div>
-          <div className="relative w-[56.25rem] h-[4.06rem]">
-            <img
-              className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <img
-              className="absolute h-[45%] w-[4%] top-[20.96%] right-[71.44%] bottom-[34.04%] left-[24.56%] max-w-full overflow-hidden max-h-full opacity-[0.65]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_420px)] w-[37.75rem] h-[2.31rem] text-[1.75rem]">
-              <div className="absolute top-[0rem] left-[20.69rem] w-[13.75rem] flex flex-row items-center justify-start gap-[0.63rem]">
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-                <div className="flex flex-row items-end justify-center gap-[0.63rem]">
-                  <i className="relative">sans</i>
-                  <div className="relative text-[0.88rem]">vs 75.0%</div>
-                </div>
-              </div>
-              <i className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_40px)] inline-block text-center w-[4.31rem] h-[2.31rem]">
-                5 : 4
-              </i>
-              <div className="absolute top-[calc(50%_-_18.5px)] left-[0rem] flex flex-row items-center justify-end gap-[0.63rem] text-right">
-                <i className="relative inline-block font-bold w-[13.75rem] h-[2.31rem] shrink-0">
-                  testNickname
-                </i>
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-              </div>
-            </div>
-            <div className="absolute top-[2.31rem] left-[49.06rem] text-[0.88rem] text-right">
-              2023.07.26 17:09:34
-            </div>
-            <div className="absolute top-[0.88rem] left-[46.69rem] text-right">
-              1 ball, speed moderate
-            </div>
-            <img
-              className="absolute top-[4rem] left-[0rem] w-[56.25rem] h-[0rem]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[0rem] left-[0.19rem]">{` `}</div>
-          </div>
-          <div className="relative w-[56.25rem] h-[4.06rem]">
-            <img
-              className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <img
-              className="absolute h-[45%] w-[4%] top-[20.96%] right-[71.44%] bottom-[34.04%] left-[24.56%] max-w-full overflow-hidden max-h-full opacity-[0.65]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_420px)] w-[37.75rem] h-[2.31rem] text-[1.75rem]">
-              <div className="absolute top-[0rem] left-[20.69rem] w-[13.75rem] flex flex-row items-center justify-start gap-[0.63rem]">
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-                <div className="flex flex-row items-end justify-center gap-[0.63rem]">
-                  <i className="relative">asdf</i>
-                  <div className="relative text-[0.88rem]">vs 75.0%</div>
-                </div>
-              </div>
-              <i className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_40px)] inline-block text-center w-[4.31rem] h-[2.31rem]">
-                5 : 4
-              </i>
-              <div className="absolute top-[calc(50%_-_18.5px)] left-[0rem] flex flex-row items-center justify-end gap-[0.63rem] text-right">
-                <i className="relative inline-block font-bold w-[13.75rem] h-[2.31rem] shrink-0">
-                  testNickname
-                </i>
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-              </div>
-            </div>
-            <div className="absolute top-[2.31rem] left-[49.06rem] text-[0.88rem] text-right">
-              2023.07.26 17:09:34
-            </div>
-            <div className="absolute top-[0.88rem] left-[46.69rem] text-right">
-              1 ball, speed moderate
-            </div>
-            <img
-              className="absolute top-[4rem] left-[0rem] w-[56.25rem] h-[0rem]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[0rem] left-[0.19rem]">{` `}</div>
-          </div>
-          <div className="relative w-[56.25rem] h-[4.06rem]">
-            <img
-              className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <img
-              className="absolute h-[45%] w-[4%] top-[20.96%] right-[71.44%] bottom-[34.04%] left-[24.56%] max-w-full overflow-hidden max-h-full opacity-[0.65]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_420px)] w-[37.75rem] h-[2.31rem] text-[1.75rem]">
-              <div className="absolute top-[0rem] left-[20.69rem] w-[13.75rem] flex flex-row items-center justify-start gap-[0.63rem]">
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-                <div className="flex flex-row items-end justify-center gap-[0.63rem]">
-                  <i className="relative">userName2</i>
-                  <div className="relative text-[0.88rem]">vs 75.0%</div>
-                </div>
-              </div>
-              <i className="absolute top-[calc(50%_-_18.5px)] left-[calc(50%_-_40px)] inline-block text-center w-[4.31rem] h-[2.31rem]">
-                5 : 4
-              </i>
-              <div className="absolute top-[calc(50%_-_18.5px)] left-[0rem] flex flex-row items-center justify-end gap-[0.63rem] text-right">
-                <i className="relative inline-block font-bold w-[13.75rem] h-[2.31rem] shrink-0">
-                  userName1
-                </i>
-                <img
-                  className="relative w-[2rem] h-[2rem] object-cover"
-                  alt=""
-                  src="/pngegg-41@2x.png"
-                />
-              </div>
-            </div>
-            <div className="absolute top-[2.31rem] left-[49.06rem] text-[0.88rem] text-right">
-              2023.07.26 17:09:34
-            </div>
-            <div className="absolute top-[0.88rem] left-[46.69rem] text-right">
-              1 ball, speed moderate
-            </div>
-            <img
-              className="absolute top-[4rem] left-[0rem] w-[56.25rem] h-[0rem]"
-              alt=""
-              src="/rectangle-12.svg"
-            />
-            <div className="absolute top-[0rem] left-[0.19rem]">custom</div>
-          </div>
-        </div>
-        <div className="absolute h-[4%] w-[15.8%] top-[32.92%] right-[79.6%] bottom-[63.08%] left-[4.6%] text-[1.25rem]">
-          <div className="absolute h-full w-[99.76%] top-[0%] right-[0%] bottom-[0%] left-[0.24%]">
-            <div className="absolute h-full w-[81.21%] top-[0%] left-[18.79%] inline-block">
-              match history
-            </div>
-            <img
-              className="absolute h-[82.49%] w-[13.47%] top-[16.05%] right-[86.53%] bottom-[1.47%] left-[0%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/vector.svg"
-            />
-          </div>
-        </div>
-        <div className="absolute top-[1.5rem] left-[1.75rem] w-[59rem] flex flex-row items-center justify-center gap-[3.75rem] text-[0.94rem]">
-          <div className="relative w-[32.44rem] h-[11.06rem] z-[2]">
-            <div className="absolute h-[68.93%] w-[77.07%] top-[31.07%] right-[0.19%] bottom-[0%] left-[22.74%]">
-              <div className="absolute h-[25.41%] w-full top-[74.59%] right-[0%] bottom-[0%] left-[0%]">
-                <div className="absolute h-[16.13%] w-full top-[0%] right-[0%] bottom-[83.87%] left-[0%]">
-                  <div className="absolute top-[0rem] left-[0rem] rounded-3xs bg-blueviolet w-[25rem] h-[0.44rem] overflow-hidden opacity-[0.2]" />
-                  <div className="absolute top-[0rem] left-[0rem] rounded-3xs bg-slateblue w-[18.63rem] h-[0.44rem] overflow-hidden" />
-                </div>
-                <div className="absolute w-[24.59%] top-[25.81%] left-[0%] leading-[150%] inline-block">
-                  winrate 75.0%
-                </div>
-              </div>
-              <div className="absolute h-[61.48%] w-full top-[0%] right-[0%] bottom-[38.52%] left-[0%] flex flex-row items-center justify-center gap-[1.94rem] text-center text-[2.5rem]">
-                <div className="relative w-[7.06rem] h-[4.69rem]">
-                  <b className="absolute h-[81.33%] w-full top-[0%] left-[0%] leading-[150%] inline-block">
-                    100
-                  </b>
-                  <div className="absolute top-[60%] left-[24.78%] text-[1.25rem] leading-[150%]">
-                    games
-                  </div>
-                </div>
-                <div className="relative w-[7.06rem] h-[4.69rem] text-midnightblue">
-                  <b className="absolute h-[81.33%] w-full top-[0%] left-[0%] leading-[150%] inline-block">
-                    75
-                  </b>
-                  <div className="absolute top-[60%] left-[31.86%] text-[1.25rem] leading-[150%]">
-                    wins
-                  </div>
-                </div>
-                <div className="relative w-[7.06rem] h-[4.69rem] text-brown">
-                  <b className="absolute h-[81.33%] w-full top-[0%] left-[0%] leading-[150%] inline-block">
-                    25
-                  </b>
-                  <div className="absolute top-[60%] left-[30.09%] text-[1.25rem] leading-[150%]">
-                    loses
-                  </div>
-                </div>
-              </div>
-            </div>
-            <img
-              className="absolute h-[6.03%] w-[76.36%] top-[25.24%] right-[0.13%] bottom-[68.74%] left-[23.51%] max-w-full overflow-hidden max-h-full"
-              alt=""
-              src="/line-12.svg"
-            />
-            <div className="absolute top-[0.25rem] left-[7.5rem] flex flex-row items-start justify-end text-[2rem]">
-              <div className="relative">testNickName</div>
-              <img
-                className="relative w-[1.25rem] h-[1.25rem] overflow-hidden shrink-0"
-                alt=""
-                src="/editbutton.svg"
-              />
-            </div>
-            <div className="absolute top-[0rem] left-[0rem] w-[6.25rem] h-[6.25rem]">
-              <img
-                className="absolute h-full w-full top-[0%] right-[0%] bottom-[0%] left-[0%] max-w-full overflow-hidden max-h-full object-cover"
-                alt=""
-                src="/pngegg-31@2x.png"
-              />
-            </div>
-          </div>
-          <img
-            className="relative w-[0.13rem] h-[9.5rem] z-[1]"
-            alt=""
-            src="/line-13.svg"
-          />
-          <div className="relative w-[15.44rem] h-[11.81rem] overflow-hidden shrink-0 z-[0] text-[1rem]">
-            <div className="absolute top-[3.94rem] left-[0rem] h-[7.88rem] overflow-y-auto flex flex-col items-start justify-start gap-[0.44rem]">
-              <div className="relative w-[15.44rem] h-[2.81rem]">
-                <div className="absolute top-[0rem] left-[0.06rem] flex flex-row items-center justify-center gap-[0.63rem]">
-                  <img
-                    className="relative w-[2rem] h-[2rem] object-cover"
-                    alt=""
-                    src="/pngegg-4@2x.png"
-                  />
-                  <div className="relative">s</div>
-                </div>
-                <img
-                  className="absolute top-[2.48rem] left-[0.02rem] w-[15.05rem] h-[0.09rem]"
-                  alt=""
-                  src="/line-10.svg"
-                />
-              </div>
-              <div className="relative w-[15.44rem] h-[2.81rem]">
-                <div className="absolute top-[0rem] left-[0.06rem] flex flex-row items-center justify-center gap-[0.63rem]">
-                  <img
-                    className="relative w-[2rem] h-[2rem] object-cover"
-                    alt=""
-                    src="/pngegg-4@2x.png"
-                  />
-                  <div className="relative">c</div>
-                </div>
-                <img
-                  className="absolute top-[2.48rem] left-[0.02rem] w-[15.05rem] h-[0.09rem]"
-                  alt=""
-                  src="/line-10.svg"
-                />
-              </div>
-              <div className="relative w-[15.44rem] h-[2.81rem]">
-                <div className="absolute top-[0rem] left-[0.06rem] flex flex-row items-center justify-center gap-[0.63rem]">
-                  <img
-                    className="relative w-[2rem] h-[2rem] object-cover"
-                    alt=""
-                    src="/pngegg-2@2x.png"
-                  />
-                  <div className="relative">r</div>
-                </div>
-                <img
-                  className="absolute top-[2.53rem] left-[0.06rem] w-[15rem] h-[0rem]"
-                  alt=""
-                  src="/rectangle-12.svg"
-                />
-              </div>
-              <div className="relative w-[15.44rem] h-[2.81rem]">
-                <div className="absolute top-[0rem] left-[0.06rem] flex flex-row items-center justify-center gap-[0.63rem]">
-                  <img
-                    className="relative w-[2rem] h-[2rem] object-cover"
-                    alt=""
-                    src="/pngegg-41@2x.png"
-                  />
-                  <div className="relative">testUserName</div>
-                </div>
-                <img
-                  className="absolute top-[2.53rem] left-[0.06rem] w-[15rem] h-[0rem]"
-                  alt=""
-                  src="/rectangle-12.svg"
-                />
-              </div>
-              <div className="relative w-[15.44rem] h-[2.81rem]">
-                <div className="absolute top-[0rem] left-[0.06rem] flex flex-row items-center justify-center gap-[0.63rem]">
-                  <img
-                    className="relative w-[2rem] h-[2rem] object-cover"
-                    alt=""
-                    src="/pngegg-41@2x.png"
-                  />
-                  <div className="relative">testUserName</div>
-                </div>
-                <img
-                  className="absolute top-[2.53rem] left-[0.06rem] w-[15rem] h-[0rem]"
-                  alt=""
-                  src="/rectangle-12.svg"
-                />
-              </div>
-              <div className="relative w-[15.44rem] h-[2.81rem]">
-                <div className="absolute top-[0rem] left-[0.06rem] flex flex-row items-center justify-center gap-[0.63rem]">
-                  <img
-                    className="relative w-[2rem] h-[2rem] object-cover"
-                    alt=""
-                    src="/pngegg-41@2x.png"
-                  />
-                  <div className="relative">testUserName</div>
-                </div>
-                <img
-                  className="absolute top-[2.53rem] left-[0.06rem] w-[15rem] h-[0rem]"
-                  alt=""
-                  src="/rectangle-12.svg"
-                />
-              </div>
-            </div>
-            <div className="absolute top-[2.44rem] left-[0.06rem] text-[0.94rem] inline-block w-[4.69rem] h-[0.88rem]">
-              result
-            </div>
-            <div className="absolute top-[0rem] left-[0rem] w-[15.44rem] h-[1.88rem]">
-              <img
-                className="absolute h-[35.56%] w-[99.06%] top-[72.22%] right-[0.84%] bottom-[-7.78%] left-[0.1%] max-w-full overflow-hidden max-h-full"
-                alt=""
-                src="/line-11.svg"
-              />
-              <div className="absolute top-[0rem] left-[0.44rem] flex flex-row items-center justify-start gap-[0.25rem]">
-                <img
-                  className="relative w-[1.5rem] h-[1.5rem] overflow-hidden shrink-0"
-                  alt=""
-                  src="/search-fill0-wght500-grad0-opsz48-1.svg"
-                />
-                <i className="relative inline-block font-bold w-[8.88rem] h-[1.19rem] shrink-0">
-                  search users
-                </i>
-              </div>
-            </div>
-          </div>
-        </div>
-        <img
-          className="absolute top-[0.63rem] right-[0.75rem] w-[2rem] h-[2rem] overflow-hidden"
-          alt=""
-          src="/cancelbutton.svg"
-        />
-      </div>
-    </div>
-  );
-};
-
-export default MyPageFrame;
